requirements.txt
# Core dependencies
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
Pillow

# NLP libraries for ATS Score Checker
spacy
nltk

# For background tasks
celery
redis

# Payment integration
razorpay
stripe

# PDF & DOCX export
xhtml2pdf
python-docx

# Structured logging
python-json-logger

# Development and testing
pytest
pytest-django
coverage
black
flake8

# Deployment
gunicorn
whitenoise
dj-database-url
python-dotenv
drf-yasg
